+++ src/main/java/kz/bars/familybudget/repository/UserRepo.java
@@ -5,9 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
+
 @Repository
 @Transactional
-public interface UserRepo extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, BigInteger> {
 
     User findByEmail(String email);
