+++ src/main/java/kz/bars/familybudget/service/BudgetService.java
@@ -3,16 +3,17 @@ package kz.bars.familybudget.service;
 import kz.bars.familybudget.dto.BudgetDto;
 import kz.bars.familybudget.model.Budget;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 
 public interface BudgetService {
 
     List<Budget> getAllBudget();
-    Budget getBudget(Long id);
+    Budget getBudget(BigInteger id);
     Budget addBudget(Budget budget);
     Budget updateBudget(Budget budget);
-    void deleteBudget(Long id);
+    void deleteBudget(BigInteger id);
     BudgetDto toDto(Budget budget);
     List<BudgetDto> getAllBudgetDto();
     List<BudgetDto> getAllBudgetBetweenDateDto(LocalDate dateFrom, LocalDate dateTo);
